% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_raw_data.R
\name{extract_bbtools}
\alias{extract_bbtools}
\title{Read information from bbtools report}
\usage{
extract_bbtools(input_dir = character())
}
\arguments{
\item{input_dir}{Path to directory with results from Juno pipeline.}
}
\value{
Tidy data frame containing information from bbtools report to be used
  for this QC report.
}
\description{
Read information from bbtools report
}
