% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_raw_data.R
\name{extract_checkm}
\alias{extract_checkm}
\title{Read information from CheckM report}
\usage{
extract_checkm(input_dir = character())
}
\arguments{
\item{input_dir}{Path to directory with results from Juno pipeline.}
}
\value{
Tidy data frame containing information from CheckM report to be used
  for this QC report
}
\description{
Read information from CheckM report
}
