% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_raw_data.R
\name{extract_quast}
\alias{extract_quast}
\title{Read information from QUAST report}
\usage{
extract_quast(input_dir = character())
}
\arguments{
\item{input_dir}{Path to directory with results from Juno pipeline.}
}
\value{
Tidy data frame containing information from QUAST report to be used
  for this QC report
}
\description{
Read information from QUAST report
}
