% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_accepted_criteria}
\alias{get_accepted_criteria}
\title{Get accepted criteria for different metrics}
\usage{
get_accepted_criteria(
  genus = character(),
  criterion = c("avg_phred", "Total length", "# contigs", "N50", "GC (\%)",
    "Average coverage", "completeness", "contamination")
)
}
\arguments{
\item{genus}{Character vector of length 1 with the name of a genus supported
by the refsamp package. Currently accepted genera: "Bordetella",
"Escherichia", "Listeria", "Neisseria", "Salmonella", "Shigella" and
"Streptococcus".}

\item{criterion}{Character vector of length 1 with one of the included
quality criteria for acceptance of a sample. The currently accepted
criteria are: "avg_phred", "Total length", "# contigs", "N50", "GC (%)",
"Average coverage", "completeness" and "contamination".}
}
\value{
Dataframe of 2-3 rows (depending on the criterion) including the
  Genus and the yintercept or the Genus and the ymin and ymax for drawing a
  geom_hline or geom_rect, depending on the criterion.These are the
  thresholds to be used for accepting samples according to a$Run_date given
  quality criterion.
}
\description{
Get accepted criteria for different metrics
}
