% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_refsamp_report}
\alias{make_refsamp_report}
\title{Make reference samples report}
\usage{
make_refsamp_report(input_dir, output_report, history_data = NULL)
}
\arguments{
\item{input_dir}{Path to directory with results from Juno pipeline.
Collection name of Juno output as assigned by iRODS in previous pipeline.
It MUST start with the date in the format "yymmdd"}

\item{output_report}{Desired file name of the output report (.html
extension).}

\item{history_data}{Path to history_data. If not provided, a history_data.csv
file will be created in the same directory where the output_report is.}
}
\value{
Html report (file name taken from output_report parameter) and a csv
  file with the history data.
}
\description{
Make reference samples report
}
