% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_raw_data.R
\name{merging_by_sample}
\alias{merging_by_sample}
\title{Merge datasets by sample}
\usage{
merging_by_sample(vector_w_dataframes, run_date)
}
\arguments{
\item{vector_w_dataframes}{Character vector with the name of the datasets to
merge. Each dataset MUST contain at least one column called "Sample" that
is common between them.}

\item{run_date}{Character vector with the date in the format "yy-mm-dd" as
given by refsamp::get_run_date()}
}
\value{
Tibble with the different datasets provided in vector_w_dataframes
  bound by the 'Sample' column.
}
\description{
Merge datasets by sample
}
