% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_raw_data.R
\name{read_tsv_report}
\alias{read_tsv_report}
\title{Read tsv report}
\usage{
read_tsv_report(file_path)
}
\arguments{
\item{file_path}{Path to tsv report.}
}
\value{
Tibble containing the tsv report
}
\description{
Read tsv report
}
