% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_time_metrics}
\alias{plot_time_metrics}
\title{Plot metric vs time}
\usage{
plot_time_metrics(dataset, metric)
}
\arguments{
\item{dataset}{Data frame to use for plot. It must contain at least one
character column named "Run_date" (yy-mm-dd) and one character column named
"Sample".}

\item{metric}{Character vector length 1 with the name of the metric to plot
(metric should be one of the column names of dataset)}
}
\value{
ggplot object containing the plot to be added
}
\description{
Plot metric vs time
}
